requirements.txt
amqp==2.4.2
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
billiard==3.6.0.0
bleach==3.1.0
celery==4.3.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-appconf==1.0.3
django-autoslug==1.9.4
django-ckeditor==5.7.1
django-countries==5.3.1
django-ffmpeg==0.0.14
django-filter==1.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-phonenumber-field==2.3.1
django-ranged-fileresponse==0.1.2
django-tinymce==2.8.0
django-twilio==0.9.3
django-user-language-middleware==0.0.2
django-watermark==0.1.8
djangorestframework==3.8.2
entrypoints==0.3
enum34==1.1.6
gunicorn==19.6.0
html5lib==1.0.1
idna==2.7
imageio==2.5.0
imageio-ffmpeg==0.3.0
ipaddress==1.0.22
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
kombu==4.5.0
Markdown==2.6.11
MarkupSafe==1.1.1
mistune==0.8.4
moviepy==1.0.0
nbconvert==5.5.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
notebook==5.7.8
numpy==1.16.3
olefile==0.45.1
packaging==17.1
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.1.0
proglog==0.1.9
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.18
Pygments==2.4.0
PyJWT==1.7.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyrsistent==0.15.2
PySocks==1.6.8
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
pyuca==1.2
pyzmq==18.0.1
redis==3.2.1
reportlab==3.5.13
requests==2.19.1
Send2Trash==1.5.0
six==1.11.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
twilio==6.25.1
urllib3==1.23
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
